# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FLAI API
  description: |
    Fully Local AI (FLAI) - API Documentation

    FLAI is a fully local personal assistant powered by artificial intelligence.

    ## Authentication
    Most endpoints require authentication via session cookie.
    Authenticate by POSTing to `/login` with credentials.

    ## Rate Limiting
    Login endpoint is rate limited: 5 requests per minute, 10 per hour.
    Send message endpoint: 15 per minute, 60 per hour.

    ## Feature Gates
    - Camera endpoints (`/api/cam/*`) require `CAMERA_ENABLED=true`
    - Debug endpoints (`/api/debug/*`, `/api/rag/*`) require `DEBUG_API_ENABLED=true`

    ## Error Responses
    All error responses return JSON with an `error` field:
    ```json
    {
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: FLAI Support
    url: https://github.com/barval/flai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://flai.example.com
    description: Production server
...
paths:
...
  /api/files/{filename}:
    get:
      tags:
        - Files
      summary: Serve uploaded file
      description: Serve uploaded files (images, audio) with session ownership validation
      security:
        - sessionAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Path in format session_id/filename
      responses:
        200:
          description: File content with appropriate MIME type
        400:
          description: Invalid filename format
        401:
          description: Not authorized
        403:
          description: Path traversal blocked or access denied
        404:
          description: File not found
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
...